+++ CHANGELOG.md
@@ -5,10 +5,21 @@ The format is based on [Keep a Changelog][keepachangelog] and this project
 adheres to [Semantic Versioning][semver].
 
 ## [Unreleased][unreleased]
+
+## [v0.10.3][v0.10.3] - 2017-10-15
+### Fixed
+- Fixed saving token file to disk. (#9)
+
+## [v0.10.2][v0.10.2] - 2017-10-05
+### Fixed
+- Fixed automatic token refreshing - thanks @bartonp! (#5)
+
 ### Changed
 - `MonzoAPI()._refresh_oath_token()` now doesn't return anything, replaces
-  current token and raises `UnableToRefreshTokenException` when token couldn't
-  be refreshed.
+  current token and raises `CantRefreshTokenError` when token couldn't be
+  refreshed.
+- Client secret is now saved in token file JSON file.
+- Cleaned up exceptions.
 
 ## [v0.10.1][v0.10.1] - 2017-09-24
 ### Fixed
@@ -38,8 +49,10 @@ adheres to [Semantic Versioning][semver].
 
 [keepachangelog]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
-[unreleased]: https://github.com/pawelad/pymonzo/compare/v0.10.1...HEAD
+[unreleased]: https://github.com/pawelad/pymonzo/compare/v0.10.2...HEAD
 [v0.9.0]: https://github.com/pawelad/pymonzo/releases/tag/v0.9.0
 [v0.9.1]: https://github.com/pawelad/pymonzo/releases/tag/v0.9.1
 [v0.10.0]: https://github.com/pawelad/pymonzo/releases/tag/v0.10.0
 [v0.10.1]: https://github.com/pawelad/pymonzo/releases/tag/v0.10.1
+[v0.10.2]: https://github.com/pawelad/pymonzo/releases/tag/v0.10.2
+[v0.10.3]: https://github.com/pawelad/pymonzo/releases/tag/v0.10.3